Support zooming to imagery's highest zoom
Essentially this extends the max zoom on a case by case basis
when an individual imagery layer is selected throught the results pane.

Fixes #200.

I had planned to also limit the panning (so the imagery can't leave the
viewport), however the default behaviour of selecting imagery doesn't
focus the map, so there's no consistent entry point for applying new
panning limits. It could be done alongside the zoom-to-fit button
behaviour, but then there is an inconsistency in panning behaviour,
which we be more confusing than helpful. So as it stands this commit
does leave a small UX problem in that you can zoom to a level where the
main map layer disappears, but only when individual imagery is selected.
As soon as imagery is deselected the default max zoom (18) is reset and
you are zoomed back out to the default max zoom if you were browsing
imagery at a higher zoom level.

Notes:
  * Imagery's metadata does not contain the max tile zoom. It needs to be
    fetched separately through the tileJSON manifest, this is somewhat
    hacky and requires a blocking XHR request. We should look to
    include at least the maxZoom level from imagery's tileJSON, or
    perhaps even include the tileJSON by default in the
    `oin-meta-generator` process.
  * Mapbox.js version is missing setMaxZoom(), consider upgrading.
  * The custom DS zoom controls are listening to `layeradd` which gets
    triggered around 150 times per map movement.
  * Commit also contains formatting updates, mainly line length.