Merged entire oam-uploader repo
This is the first step in the merge. Everything works but there
is some redundancy and minor TODOs. The main issue is the duplication
of CSS, see: styles/uploader/*. Consolidating the CSS will also require
bringing the uploader design in line with the browser.

I did manage to fairly easily consolidate the collecticons, but there
were 2 missing icons oam-design-system for dropbox and gdrive, see:
styles/uploader/_buttons.scss.

The merge includes commit history.

Touches #220
Closes hotosm/oam-uploader#76