Use Windows 8 instead of Windows 10
This was extraorindarily annoying to discover. Suddenly, Browserstack
tests stopped connecting through the local tunnel. And I tried
everything. Eventually I discovered that it was to do with the OS
version. My thinking is that I had done some Browserstack Live testing
which requires installing a plugin, onto the remote Browserstackk
machine, which I suspect overrides the local Browserstack binary setup.
Which is to say, you can't use both the Browserstack binary for Automate
and the plugin for Live. Not to mention that this must mean that
individual test sessiosn don't run on idempotent VMs.